+++ examples/data-sources/fmc_as_paths/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_as_paths" "example" {
+  items = {
+    "my_as_paths" = {
+    }
+  }
+}