+++ internal/provider/data_source_fmc_as_paths_test.go
@@ -1,0 +1,81 @@
+// Copyright © 2023 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccDataSource
+
+func TestAccDataSourceFmcASPaths(t *testing.T) {
+	var checks []resource.TestCheckFunc
+	checks = append(checks, resource.TestCheckResourceAttrSet("data.fmc_as_paths.test", "items.my_as_paths.id"))
+	checks = append(checks, resource.TestCheckResourceAttrSet("data.fmc_as_paths.test", "items.my_as_paths.type"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_as_paths.test", "items.my_as_paths.overridable", "false"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_as_paths.test", "items.my_as_paths.entries.0.action", "PERMIT"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_as_paths.test", "items.my_as_paths.entries.0.regular_expression", "^(100|200)$"))
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		ErrorCheck:               func(err error) error { return testAccErrorCheck(t, err) },
+		Steps: []resource.TestStep{
+			{
+				Config: testAccDataSourceFmcASPathsConfig(),
+				Check:  resource.ComposeTestCheckFunc(checks...),
+			},
+		},
+	})
+}
+
+// End of section. //template:end testAccDataSource
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testPrerequisites
+// End of section. //template:end testPrerequisites
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccDataSourceConfig
+
+func testAccDataSourceFmcASPathsConfig() string {
+	config := `resource "fmc_as_paths" "test" {` + "\n"
+	config += `	items = { "my_as_paths" = {` + "\n"
+	config += `		overridable = false` + "\n"
+	config += `		entries = [{` + "\n"
+	config += `			action = "PERMIT"` + "\n"
+	config += `			regular_expression = "^(100|200)$"` + "\n"
+	config += `		}]` + "\n"
+	config += `	}}` + "\n"
+	config += `}` + "\n"
+
+	config += `
+		data "fmc_as_paths" "test" {
+			depends_on = [fmc_as_paths.test]
+			items = {
+				"my_as_paths" = {
+				}
+			}
+		}
+	`
+	return config
+}
+
+// End of section. //template:end testAccDataSourceConfig